% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullwhip.R
\name{bullwhip-deprecated}
\alias{bullwhip-deprecated}
\title{Bullwhip effect}
\usage{
bullwhip(method, phi, L, p, alpha)
}
\arguments{
\item{method}{Character string specifing which method to use}

\item{phi}{A vector of autoregressive parameters}

\item{L}{A positive lead-time}

\item{p}{Order to be used in the SMA method}

\item{alpha}{Smoothing factor to be used in the ES method (0 < alpha < 1)}
}
\value{
The measure for the bullwhip effect
}
\description{
\code{bullwhip} computes the increase of demand
variability for a simple two-stage supply chains
consisting of a single retailer and a single manufacturer using three
forcasting methods: Minimum Mean Square Error (MMSE), Simple Moving
Average (SMA) and Exponential Smoothing (ES) when the demand follows
a known stationary AR(1) stochastic process.
}
\details{
The \code{bullwhip} function has been deprecated and will be made defunct; use 
the bullwhipgame package.
}
\examples{

\dontrun{

bullwhip("SMA",0.9,2,4)

bullwhip("ES",0.9,2,0,0.6)

bullwhip("MMSE",0.9,2) 
}

}
\seealso{
\code{\link{pkgSCperf-deprecated}}
}
\keyword{internal}
