% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCperf.R, R/SSL.R, R/bullwhip.R,
%   R/pkgSCperf-deprecated.R
\name{SCperf}
\alias{SCperf}
\alias{SSL}
\alias{bullwhip}
\alias{pkgSCperf-deprecated}
\title{Deprecated functions in the \pkg{SCperf} package}
\usage{
SCperf(phi, theta, L = L, SL = 0.95)

SSL(method = c("MMSE", "SMA", "ES"), phi, L, p, alpha, SL)

bullwhip(method = c("MMSE", "SMA", "ES"), phi, L, p, alpha)
}
\description{
The functions listed below are provided for compatibility 
with older version of the \pkg{SCperf} package. They have been deprecated 
and will be defunct in the next version of this package.
Help pages for deprecated functions are available at 
\code{help("-deprecated")}
}
\details{
The \code{SCperf} function has been deprecated and will be made defunct; use 
the bullwhipgame package.

The \code{SSL} function has been deprecated and will be made defunct; use 
the bullwhipgame package.

The \code{bullwhip} function has been deprecated and will be made defunct; use 
the bullwhipgame package.
}
\keyword{internal}
