% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROP.R
\name{ROP}
\alias{ROP}
\title{Reorder Point}
\usage{
ROP(SL, md, sd, L = 1)
}
\arguments{
\item{SL}{Service level, a value between 0 and 1}

\item{md}{Mean demand}

\item{sd}{Standart deviation of the demand}

\item{L}{A positive lead-time}
}
\value{
Reorder point level
}
\description{
\code{ROP} computes the reorder point level of inventory. When inventory falls to this amount, a new order must be made.
}
\examples{

 ROP(0.9,2500,500,6)

}
\seealso{
\code{\link{SS}}
}
