% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCperf.R
\name{SCperf-deprecated}
\alias{SCperf-deprecated}
\title{Supply Chain Performance}
\usage{
SCperf(phi, theta, L, SL)
}
\arguments{
\item{phi}{A vector of autoregressive parameters}

\item{theta}{A vector of moving-average parameters}

\item{L}{A positive lead-time}

\item{SL}{Service level, (default:0.95)}
}
\value{
Analytical measure for the bullwhip effect
}
\description{
Computes the bullwhip effect for an stationary ARMA(p,q) demand process.
}
\details{
The \code{SCperf} function has been deprecated and will be made defunct; use 
the bullwhipgame package.
}
\examples{

\dontrun{

#ARMA(1,1) case

SCperf(phi=0.95,theta=0.1,L=2,SL=0.99)

#AR(2) case

SCperf(phi=c(0.8,-0.2),theta=0,L=1)
}

}
\references{
- Zhang, X. (2004b). Evolution of ARMA demand in supply
  chains. Manufacturing and Services Operations Management, 6 (2), 195-198.

- Silva Marchena, M. (2010) Measuring and implementing the bullwhip effect under
  a generalized demand process. \url{ http://arxiv.org/abs/1009.3977}
}
\seealso{
\code{\link{pkgSCperf-deprecated}}
}
\keyword{internal}
