% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSL.R
\name{SSL-deprecated}
\alias{SSL-deprecated}
\title{Safety stock over lead-time}
\usage{
SSL(method, phi, L, p, alpha, SL)
}
\arguments{
\item{method}{Character string specifing which method to use}

\item{phi}{A vector of autoregressive parameters}

\item{L}{A positive lead-time}

\item{p}{Order to be used in the SMA method}

\item{alpha}{Smoothing factor to be used in the ES method (0 < alpha < 1)}

\item{SL}{Service level}
}
\value{
Safety stock level over lead-time
}
\description{
\code{SSL} computes the safety stock level over lead-time for three forecasting
methods: Minimum Mean Square Error (MMSE), Simple Moving Average (SMA) and
Exponential Smoothing (ES) when the demand follows a stationary AR(1) stochastic process.
}
\details{
The \code{SSL} function has been deprecated and will be made defunct; use 
the bullwhipgame package.
}
\examples{

\dontrun{

SSL("MMSE",0.15,2,4,0.7,0.95)

SSL("SMA",0.15,2,4,0.7,0.95)

SSL("ES",0.15,2,4,0.7,0.95) 
}

}
\seealso{
\code{\link{pkgSCperf-deprecated}}
}
\keyword{internal}
